[ISSUE] #2 Further changes due to renaming the controller
