+++ src/main/kotlin/com/schulverwaltung/controller/interfaces/ISecretController.kt
@@ -1,4 +1,4 @@
-package com.schulverwaltung.controllers.interfaces
+package com.schulverwaltung.controller.interfaces
 
 import com.schulverwaltung.dto.interfaces.ISecret
